# public/swagger.yaml
openapi: 3.0.0
info:
  title: 'Therapeutic Radiesthesia API'
  description: '`API para demonstrar endpoints básicos do sistema`<br><ul><li>Essa documentação está sendo desenvolvida gradualmente, todos os endpoints estão passando por uma revisão.</li><li>Os endpoints que precisam estar com a autorização estão com um cadeado indicando o uso.</li><li>Para gerar o token, utilize o endpoint login.</li><li>Endpoints marcados como acesso ADMIN.</li></ul>'
  contact:
    name: 'Paulo Henrique'
    url: 'https://api.conect.app'
    email: webmaster@multidesk.io
  version: 1.0.0
servers:
  -
    url: 'https://api.multidesk.io'
    description: 'Servidor online'
  -
    url: 'http://localhost:8000'
    description: 'Servidor local'
paths:
...
  /api/v1/user/me:
    get:
      tags:
        - Usuários
      summary: 'Obter informações do usuário autenticado'
      description: 'Retorna as informações do usuário autenticado usando JWT'
      operationId: c01da61cb97dc778549fec55810d4b26
      responses:
        '200':
          description: 'Informações do usuário'
          content:
            application/json:
              schema:
                properties:
                  id: { type: integer, example: 1 }
                  name: { type: string, example: 'John Doe' }
                  email: { type: string, example: john.doe@example.com }
                  photo: { type: string, example: 'https://example.com/photo.jpg' }
                  role: { type: string, example: PROFISSIONAL }
                  plan: { type: string, example: Premium }
                  lang: { type: string, example: en }
                  languages: { type: array, items: { type: string }, example: 'en, es' }
                  description: { type: string, example: 'Experienced professional in healthcare.' }
                  education: { type: string, example: 'PhD in Radiology' }
                  department: { type: string, example: Radiology }
                  social_networks: { type: object, example: 'https://linkedin.com/in/johndoe, https://instagram.com/in/johndoe' }
                  company: { type: string, example: 'Medical Inc.' }
                  birthdate: { type: string, format: date, example: '1985-08-15' }
                  show_personal_chart_dashboard: { type: boolean, example: true }
                  show_family_chart_dashboard: { type: boolean, example: false }
                  show_friends_chart_dashboard: { type: boolean, example: true }
                  show_appointments_chart_dashboard: { type: boolean, example: true }
                  show_basic_info_dashboard: { type: boolean, example: true }
                  receive_updates_email: { type: boolean, example: true }
                  receive_updates_sms: { type: boolean, example: false }
                  receive_updates_whatsapp: { type: boolean, example: true }
                  receive_scheduling_reminders: { type: boolean, example: true }
                  receive_cancellation_reminders: { type: boolean, example: true }
                  statistics: { type: object, example: '{"total_appointments": 50, "cancelled_appointments": 5}' }
                type: object
        '401':
          description: 'Não autorizado'
      security:
        -
          bearerAuth: []
    put:
      tags:
        - Usuários
      summary: 'Atualizar informações do usuário autenticado'
      description: 'Permite que o usuário autenticado atualize suas informações de perfil.'
      operationId: 21ad00e8e69df5d13e16c2a1c9c190e8
      requestBody:
        required: true
        content:
          application/json:
            schema:
              required:
                - name
                - photo
                - phone
                - lang
                - languages
                - description
                - education
                - department
                - company
                - birthdate
              properties:
                name:
                  type: string
                  example: 'TERAPEUTA TP'
                photo:
                  type: string
                  example: 'https://vinha-dev.s3.us-east-1.amazonaws.com/admin_geral/1/66c4784ed3f9f.png'
                phone:
                  type: string
                  example: '+55 62 9 9999 9999'
                lang:
                  type: string
                  example: pt-BR
                languages:
                  type: string
                  example: 'En, Es, pt-BR'
                description:
                  type: string
                  example: 'Coloque o texto que quiser'
                education:
                  type: string
                  example: FACULDADE
                department:
                  type: string
                  example: Radiology
                social_networks:
                  type: string
                  example: social_networks
                company:
                  type: string
                  example: company
                birthdate:
                  type: string
                  format: date
                  example: '1990-01-01'
                show_personal_chart_dashboard:
                  type: boolean
                  example: true
                show_family_chart_dashboard:
                  type: boolean
                  example: true
                show_friends_chart_dashboard:
                  type: boolean
                  example: true
                show_appointments_chart_dashboard:
                  type: boolean
                  example: true
                show_basic_info_dashboard:
                  type: boolean
                  example: true
                receive_updates_email:
                  type: boolean
                  example: true
                receive_updates_sms:
                  type: boolean
                  example: true
                receive_updates_whatsapp:
                  type: boolean
                  example: true
                receive_scheduling_reminders:
                  type: boolean
                  example: true
                receive_cancellation_reminders:
                  type: boolean
                  example: true
              type: object
      responses:
        '200':
          description: 'Perfil atualizado com sucesso'
          content:
            application/json:
              schema:
                properties:
                  message: { type: string, example: 'Perfil atualizado com sucesso' }
                  data: { type: object, example: '{"id": 1, "name": "TERAPEUTA TP", "email": "user@example.com"}' }
                type: object
        '400':
          description: 'Erro de validação'
          content:
            application/json:
              schema:
                properties:
                  error: { type: string, example: 'Nome e foto são obrigatórios' }
                type: object
        '500':
          description: 'Erro inesperado no servidor'
          content:
            application/json:
              schema:
                properties:
                  error: { type: string, example: 'Internal server error' }
                type: object
      security:
        -
          bearerAuth: []
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      bearerFormat: JWT
      scheme: bearer
...